refactor: constants.ts를 BoardEmpty.constants.ts, DayBar.constants.ts로 이름 변경
