+++ src/components/menu/MenuBoard/_BoardEmpty/BoardEmpty.constants.ts
